requirement.txt
google-adk
python-dotenv
google-generativeai
psutil
litellm
yfinance
